@@ -3,20 +3,24 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
+
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.11-SNAPSHOT</version>
+		<version>3.1.3-SNAPSHOT</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
+
 	<groupId>com.codespark</groupId>
 	<artifactId>spring-boot-rest-hateoas</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>spring-boot-rest-hateoas</name>
 	<description>Building a Hypermedia-Driven RESTful Web Service</description>
+	
 	<properties>
-		<java.version>11</java.version>
+		<java.version>17</java.version>
 	</properties>
+	
 	<dependencies>
 		<!-- For adding HATEOAS + web support -->
 		<dependency>
@@ -65,6 +69,7 @@
 			</plugin>
 		</plugins>
 	</build>
+	
 	<repositories>
 		<repository>
 			<id>spring-milestones</id>
